raspberrypi/requirements.txt
requests==2.31.0
RPi.GPIO==0.7.1
pygame==2.5.2
gTTS==2.4.0
python-dotenv==1.0.0 